Average.Time(SizeSpectra)	R Documentation

A Time Averaging Function

Description

A function to average the output from several timesteps given a time range

Usage


Average.Time(species,time.lim)

Arguments

species 	 an object of class species.results

time.lim 	 a vector of length 2 specifying the start and end times of the range to be averaged over

Details

~~ If necessary, more details than the description above ~~

Value

Returns an object of class timestep.datawhich has six slots;

data 	 a data frame containg the a timestep value of the averaged results

spatial.dim 	 an integer specifying the spatial dimension of the results

trange 	 a numeric vector containing all of the t values outputed

mrange 	 a numeric vector containing all of the m values outputed

xrange 	 a numeric vector containing all of the x values outputed

yrange 	 a numeric vector containing all of the y values outputed

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

Matt Castle

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples




[Package SizeSpectra version 1.0 Index]