Read.In(SizeSpectra)	R Documentation

The Data Reading Function

Description

This reads in the output from the text files produced by the function SizeSpectrum

Usage


Read.In(filename)

Arguments

filename 	 A full filename and path to a text file produced by the function SizeSpectrum

Details

~~ If necessary, more details than the description above ~~

Value

Returns an object of class species.results which has 8 slots containing all of the information used to produce the output file. These are

results 	 a data frame containing the actual species results in a standardised format

run 	 a data frame containing the values and strings used in the model run

species 	 an object of class species.params containing all the parameters used to generate the species results

trange 	 a numeric vector containing all of the t values outputed

mrange 	 a numeric vector containing all of the m values outputed

xrange 	 a numeric vector containing all of the x values outputed

yrange 	 a numeric vector containing all of the y values outputed

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

Matt Castle

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples




[Package SizeSpectra version 1.0 Index]