Setup.Grid(SizeSpectra)	R Documentation

Grid Setup Function

Description

Produces a list of grid discretisation values in a standardised format.

Usage


Setup.Grid(mmin, m1, mmax, mstep, mout, t1, tmax, tstep, toutmin, toutmax, toutstep, xmin, xmax, xstep, xout, ymin, ymax, ystep, yout)

Arguments

mmin 	 minimum mass value for entire size spectra (in log(g))

m1 	 maximum mass value for plankton size spectra (in log(g))

mmax 	 maximum mass value for entire size spectra (in log(g))

mstep 	 mass step size used in finite differncing calculation. This must be an exact divisor of the total mass range (mmax-mmin)

mout 	 mass step size that is outputed to results text file. This must be a multiple of the mstep value (but not necessarily a divisor of the total mass range)

t1 	 time value at which spatial movement starts (in days). (All calculations start at a time value of 0)

tmax 	 maximum time value to calculate to (in days)

tstep 	 time step size used in finite differencing calculation. This must be an exact divisor of the total time range (tmax)

toutmin 	 minimum time value that is outputed to results text file

toutmax 	 maximum time value that is outputed to results text file

toutstep 	 time step size that is outputed to results text file

xmin 	 minimum x-space value (in m)

xmax 	 maximum x-space value (in m)

xstep 	 x-space step size used in finite differencing calculation. This must be an exact divisor of the total x-space range (xmax-xmin)

xout 	 x-space step size that is outputed to results text file

ymin 	 minimum y-space value (in m)

ymax 	 maximum y-space value (in m)

ystep 	 y-space step size used in finite differencing calculation. This must be an exact divisor of the total y-space range (ymax-ymin)

yout 	 y-space step size that is outputed to results text file

Details

~~ If necessary, more details than the description above ~~

Value

Returns an object of class grid.params containing the input arguments as slots either as specified or containing default values.

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

Matt Castle

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples




[Package SizeSpectra version 1.0 Index]