Setup.Initial(SizeSpectra)	R Documentation

Initial Values Setup Function

Description

Produces a text file containing the initial values for a given species

Usage


Setup.Initial(grid.in,mfunc,xfunc,yfunc)

Arguments

grid.in 	 minimum mass value for entire size spectra (in log(g))

mfunc 	 maximum mass value for plankton size spectra (in log(g))

xfunc 	 maximum mass value for entire size spectra (in log(g))

yfunc 	 mass step size used in finite differncing calculation. This must be an exact divisor of the total mass range (mmax-mmin)

Details

~~ If necessary, more details than the description above ~~

Value

Returns the filename of the text file produced.

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

Matt Castle

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples




[Package SizeSpectra version 1.0 Index]