Setup.Run(SizeSpectra)	R Documentation

Run Setup Function

Description

Produces a list of the main run parameters for the model under investigation

Usage


Setup.Run(no.species, spatial.dim, coupled.flag, initial.flag, filename)

Arguments

no.species 	 an integer value for the number of pelagic species in the model

spatial.dim 	 a value determining the spatial dimension of the model. Either 0, 1 or 2

coupled.flag 	 a logical value specifying whether the species are to interact

initial.flag 	 a logical value specifying whether the initial distributions are to be inputed by the user

filename 	 a root filename for the output text files

Details

Value

Returns an object of class run.params containing the input arguments as slots.

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

Matt Castle

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples




[Package SizeSpectra version 1.0 Index]