Setup.Species(SizeSpectra)	R Documentation

Species Setup Function

Description

Produces a list of species specific constants in a standardised format.

Usage


Setup.Species(alpha, beta, lambda, A, K, mu_0, mu_s, epsilon, u_0, sig, q_0, prey, pred, comp, gamma_prey, gamma_pred, gamma_comp, speciesname)

Arguments

alpha 	 The volume search allometric exponent

beta 	 The intrinsic mortality allometric exponent

lambda 	 The slope of the default plankton size spectrum

A 	 The volume search coefficient

K 	 The energy conversion efficiency

mu_0 	 The intrinsic mortality coefficient

mu_s 	 The senescence mortality coefficient

epsilon 	 The senescence mortality 'sharpness' coefficient

u_0 	 The intercept of the default plankton size spectrum

sig 	 The feeding mass ratio kernel standard deviation

q_0 	 The feeding mass ratio kernel mean

prey 	 The prey seeking spatial movement coefficient

pred 	 The predator avoiding spatial movement coefficient

comp 	 The competition avoiding spatial movement coefficient

gamma_prey 	 The prey seeking spatial movement allometric exponent

gamma_pred 	 The predator avoiding spatial movement allometric exponent

gamma_comp 	 The competition avoiding spatial movement allometric exponent

speciesname 	 The name of the species. This must be unique if more than one species is under investigation as it used to construct the output results text files.

Details

~~ If necessary, more details than the description above ~~

Value

Returns an object of class species.params list containing the arguments either as specified or containing default values.

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

Matt Castle

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples




[Package SizeSpectra version 1.0 Index]