SizeSpectra(SizeSpectra)	R Documentation

Marine Size-Spectra Package

Description

Spatial Size-spectra analysis functions for aquatic ecosystems.

Details

Package: 	 SizeSpectra

Type: 	 Package

Version: 	 1.0

Date: 	 2007-07-11

License: 	 What license is it under?

This package contains various functinos for creating plotting and analysing spatially explicit predation driven marine size spectra. It can calculate and analyse size spectra in 0, 1 and 2 spatial dimensions, and for any number of distinct species. itemSetup.Grid itemSetup.Run itemSetup.Species itemSetup.Initial itemSizeSpectrum itemRead.In itemExtract.Time itemAverage.Time itemPlot.Spectrum itemAnimate

Author(s)

Matt castle

Maintainer: <mattcastle@cefas.co.uk>

References

~~ Literature or other references for background information ~~

See Also

~~ Optional links to other man pages, e.g. ~~ ~~ <pkg> ~~

Examples




[Package SizeSpectra version 1.0 Index]