SizeSpectrum(SizeSpectra)	R Documentation

Dynamic SizeSpectrum Calculation Function

Description

Outputs 3 text files per species into the working directory. One containing the actual number densities, one containg the groth values and one containing the mortality values.

Usage


SizeSpectrum(run.in, grid.in, species.in, sp.initial)

Arguments

run.in 	 An object of class run.params

grid.in 	 An object of class grid.params

species.in 	 A list of objects of class species.params, one for each species to be included

sp.initial 	 An optional list of filenames containing the initial distributions of the species to be included

Details

This function makes a call to C code to perform the dynamic size spectrum calculation, using the values specified in the input arguments.

Value

Returns a list of txt files containing the output produced.

Warning

....

Note

~~further notes~~

~Make other sections like Warning with section{Warning }{....} ~

Author(s)

Matt Castle

References

~put references to the literature/web site here ~

See Also

~~objects to See Also as help, ~~~

Examples




[Package SizeSpectra version 1.0 Index]